skills/openclaw-webhooks/examples/fastapi/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.24.0
